requirements.txt
streamlit
crewai
crewai_tools
python-dotenv
langchain_openai
langchain_community